Matlab Generator for 2D Images
These codes will create a white image with black particles given an
initial volume fraction of particles. The aspect ratio of particles can
be defined. The orientation of particles can be defined. Both these
parameters can be assigned distributions as well. The particles are
then placed via an implementation of the nearest neighbor algorithm.